Linux compile fix.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@14632 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ video/VideoReader.h
@@ -19,9 +19,9 @@ struct FrameInfo
 	
 	double frame_time;
 	uint8* frame_buffer;
-	size_t width;
-	size_t height;
-	size_t stride_B; // >= 0
+	uint64 width;
+	uint64 height;
+	uint64 stride_B; // >= 0
 	bool top_down;
 
 	void* media_buffer;
